+++ HISTORY.md
@@ -2,6 +2,8 @@
 
 # Unpublished changes since 14.4.0
 
+- Fix: #3450 support multiplication of valueless units by arbitrary types
+  (#3454).
 - Feat: increase performance of the `map` and `forEach` methods of 
   `DenseMatrix` (#3446). Thanks @dvd101x.